requirements.txt
Flask==1.1.2
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.2
Pillow==9.3.0
Jinja2==2.11.3
MarkupSafe==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0